Refactor all log.Fatal calls to log.Panic to prevent issues with defer
