+++ cmd/plugin_init.go
@@ -22,7 +22,7 @@ var pluginInitCmd = &cobra.Command{
 				AttachStacktrace: config.DefaultAttachStacktrace,
 			})
 			if err != nil {
-				log.Fatal("Sentry initialization failed: ", err)
+				log.Panic("Sentry initialization failed: ", err)
 			}
 
 			// Flush buffered events before the program terminates.
